requirements.txt
langchain==0.2.6
langchain-openai==0.1.14
langchain_community==0.2.6
python-dotenv==1.0.1
anthropic==0.30.1
langchain-anthropic==0.1.19
chromadb==0.5.3
openai==1.35.10
sentence_transformers==2.2.2
InstructorEmbedding==1.0.1
flask==2.3.3
flask-wtf==1.2.1